module_lattice

Module algebra

Linear algebra with degree-256 polynomials over a prime-order field, vectors of such polynomials, and NTT polynomials / vectors.

Structs

Elem

An Elem is a member of the specified prime-order field.

NttMatrix

A K x L matrix of NTT-domain polynomials.

NttPolynomial

An NttPolynomial is a member of the NTT algebra T_q = Z_q[X]^256 of 256-tuples of field elements.

NttVector

An NttVector is a vector of polynomials from T_q of length K.

Polynomial

A Polynomial is a member of the ring R_q = Z_q[X] / (X^256) of degree-256 polynomials over the finite field with prime order q.

Vector

A Vector is a vector of polynomials from R_q of length K.

Traits

Field

Finite field with efficient modular reduction for lattice-based cryptography.

MultiplyNtt

Perform multiplication in the NTT domain.