module_lattice

Struct NttPolynomial

pub struct NttPolynomial<F: Field>(pub Array<Elem<F>, U256>);

An NttPolynomial is a member of the NTT algebra T_q = Z_q[X]^256 of 256-tuples of field elements.

NTT polynomials can be added and subtracted, negated, and multiplied by scalars. We do not define multiplication of NTT polynomials here: that is defined by the downstream crate using the MultiplyNtt trait.

We also do not define the mappings between normal polynomials and NTT polynomials (i.e., between R_q and T_q).

Tuple Fields
0: Array<Elem<F>, U256>
Implementations

impl<F: Field> NttPolynomial<F>

pub const fn new(x: Array<Elem<F>, U256>) -> Self

Create a new NTT polynomial.